requirements/base.txt
# Latest Django 1.5.x
django==1.5.4

# Configuration
django-configurations==0.5.1
django-secure==1.0
django-cache-url==0.6.0
dj-database-url==0.2.2

# Forms
django-braces==1.2.2
django-crispy-forms==1.4.0
django-floppyforms==1.1

# Models
South==0.8.2
django-model-utils==1.5.0

# images
Pillow==2.1.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.13.0

# For the persistance stores
psycopg2==2.5.1

# Unicode slugification
unicode-slugify==0.1.1
django-autoslug==1.7.1  

# Useful things
django-avatar==2.0

# Your custom requirements go here
dj-static==0.0.5
django-postman==3.0.1
django-notification==1.1.1
django-taggit==0.10
django-select2==4.2.1

# TODO: evaluate these packages for use to populate the Location field
#django-cities-light==2.1.0
# an alternative to django-cities-lite
#djcities==0.1

# not using anymore
#django-select2-forms
#django-chosen==0.1